# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Wine App API
  version: 0.1.0
  description: Current API endpoints for health, authentication, wines, reviews, stats, and wine photos. Web clients continue to use the Symfony session cookie; mobile clients should prefer bearer JWT auth.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://dev-api.example.com
    description: Shared development environment placeholder
  - url: https://staging-api.example.com
    description: Staging environment placeholder
  - url: https://api.example.com
    description: Production environment placeholder
paths:
...
  /api/dos:
    post:
      summary: Create denomination of origin
      operationId: postDo
      tags: [DO]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDoRequest'
      responses:
        '201':
          description: DO created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDoResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: DO already exists for the same country and name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List all denominations of origin
      operationId: getDosList
      tags: [DO]
      parameters:
        - in: query
          name: name
          description: Filter by DO name (case-insensitive contains)
          schema:
            type: string
        - in: query
          name: country
          description: Filter by country
          schema:
            type: string
            enum: [spain, france, italy, portugal, germany, argentina, chile, united_states, south_africa, australia]
        - in: query
          name: region
          description: Filter by DO region (case-insensitive contains)
          schema:
            type: string
        - in: query
          name: user_ids
          description: >
            Comma-separated user IDs. Returns only DOs where, within the same DO, at least one wine has been reviewed by every specified user.
          schema:
            type: string
            example: '1,2'
        - in: query
          name: sort_by_1
          description: Primary sort field. Allowed values are `country`, `region`, `name`.
          schema:
            type: string
            enum: [country, region, name]
            default: country
        - in: query
          name: sort_by_2
          description: Secondary sort field. Allowed values are `country`, `region`, `name`.
          schema:
            type: string
            enum: [country, region, name]
            default: region
        - in: query
          name: sort_by_3
          description: Tertiary sort field. Allowed values are `country`, `region`, `name`.
          schema:
            type: string
            enum: [country, region, name]
            default: name
      responses:
        '200':
          description: Denominations of origin list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoListResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/dos/{id}:
    put:
      summary: Partially update denomination of origin by id
      operationId: putDoById
      tags: [DO]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDoRequest'
      responses:
        '204':
          description: DO updated successfully
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: DO not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete denomination of origin by id
      operationId: deleteDoById
      tags: [DO]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: DO deleted successfully
        '404':
          description: DO not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: DO has associated wines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: PHPSESSID
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Invalid credentials.
...
    UpdateDoRequest:
      type: object
      properties:
        name:
          type: string
        region:
          type: string
        country:
          type: string
          enum: [spain, france, italy, portugal, germany, argentina, chile, united_states, south_africa, australia]
        country_code:
          type: string
          minLength: 2
          maxLength: 2
          nullable: true
        do_logo:
          type: string
          nullable: true
          description: Logo filename or path using an image extension (`jpg`, `jpeg`, `png`, `webp`, `gif`, `avif`).
        map_data:
          allOf:
            - $ref: '#/components/schemas/DoMapData'
          nullable: true
    CreateDoRequest:
      type: object
      required: [name, region, country, country_code]
      properties:
        name:
          type: string
        region:
          type: string
        country:
          type: string
          enum: [spain, france, italy, portugal, germany, argentina, chile, united_states, south_africa, australia]
        country_code:
          type: string
          minLength: 2
          maxLength: 2
        do_logo:
          type: string
          nullable: true
          description: Logo filename or path using an image extension (`jpg`, `jpeg`, `png`, `webp`, `gif`, `avif`).
        map_data:
          allOf:
            - $ref: '#/components/schemas/DoMapData'
          nullable: true
    CreateDoResponse:
      type: object
      required: [do]
      properties:
        do:
          type: object
          required: [id]
          properties:
            id:
              type: integer
              format: int64
...
    DoListItem:
      type: object
      required: [id, name, region, country, country_code, do_logo, region_logo, map_data]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        region:
          type: string
        country:
          type: string
          enum: [spain, france, italy, portugal, germany, argentina, chile, united_states, south_africa, australia]
        country_code:
          type: string
          minLength: 2
          maxLength: 2
        do_logo:
          type: string
          nullable: true
          example: rioja_DO.png
        region_logo:
          type: string
          nullable: true
          example: la_rioja.png
        map_data:
          allOf:
            - $ref: '#/components/schemas/DoMapData'
          nullable: true
    DoMapData:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: float
          minimum: -90
          maximum: 90
        lng:
          type: number
          format: float
          minimum: -180
          maximum: 180
        zoom:
          type: integer
          minimum: 1
          maximum: 18
...
    DoListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DoListItem'